refactor(tui): split scrolling into a separate mod from event list
